+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.0.2
 
 * Also accept `django.db.backends.postgresql` alias as Postgresql driver (#1).
+* Do not serialize `weeks` value when storing in the database, which caused those days to be counted twice (#2).
 
 ## v1.0.1